# Add RegexFindOptimization for embedded strings