… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexFindOptimizations.cs
… src/libraries/System.Text.RegularExpressions/tests/UnitTests/RegexFindOptimizationsTests.cs