Currently if a pattern begins with a multi-character string, we'll use IndexOf with that substring to find the next possible match location.  But if that multi-character string is a non-zero fixed distance into the pattern, we won't see it.  This changes that, letting us find such strings based on the fixed-distance sets we're already gathering, and using IndexOf for it.

From dotnet/performance:

| Method |         Toolchain |    Pattern |         Options |         Mean | Ratio |
|------- |------------------ |----------- |---------------- |-------------:|------:|
|  Count | \main\corerun.exe | [a-z]shing |            None |     9.699 ms |  1.00 |
|  Count |   \pr\corerun.exe | [a-z]shing |            None |     2.206 ms |  0.23 |
|        |                   |            |                 |              |       |
|  Count | \main\corerun.exe | [a-z]shing |        Compiled |     7.210 ms |  1.00 |
|  Count |   \pr\corerun.exe | [a-z]shing |        Compiled |     2.155 ms |  0.30 |
|        |                   |            |                 |              |       |
|  Count | \main\corerun.exe | [a-z]shing | NonBacktracking |     9.946 ms |  1.00 |
|  Count |   \pr\corerun.exe | [a-z]shing | NonBacktracking |     2.352 ms |  0.24 |

There are ~170 patterns in our corpus that benefit from this.