@AndyAyersMS 2022-04-22
Improvements on ubuntu x64: https://github.com/dotnet/perf-autofiling-issues/issues/4747